requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
idna==3.3
importlib-metadata==4.11.1
libretranslatepy==2.1.1
lxml==4.8.0
mysql-connector-python==8.0.28
protobuf==3.19.4
requests==2.27.1
soupsieve==2.3.1
translate==3.6.1
typing-extensions==4.1.1
urllib3==1.26.8
zipp==3.7.0
